# swagger/swagger.json
{"openapi":"3.0.0","paths":{"/countries":{"post":{"operationId":"CountryController_create","summary":"Создать страну с валютами","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCountryDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCountryResponse"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationErrorResponse"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerErrorResponse"}}}}},"tags":["Countries"]},"get":{"operationId":"CountryController_findAll","summary":"Получить список стран и валют","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetCountriesResponse"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerErrorResponse"}}}}},"tags":["Countries"]}},"/countries/{id}/status":{"patch":{"operationId":"CountryController_updateStatus","summary":"Изменить активность страны","parameters":[{"name":"id","required":true,"in":"path","description":"ID страны","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStatusDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCountryStatusResponse"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationErrorResponse"}}}},"403":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForbiddenErrorResponse"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundErrorResponse"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerErrorResponse"}}}}},"tags":["Countries"]}},"/countries/{countryId}/currency/{code}/status":{"patch":{"operationId":"CountryController_updateCurrencyStatus","summary":"Изменить активность валюты","parameters":[{"name":"countryId","required":true,"in":"path","description":"ID страны","schema":{"type":"string"}},{"name":"code","required":true,"in":"path","description":"Код валюты ISO","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStatusDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCurrencyStatusResponse"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotFoundErrorResponse"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerErrorResponse"}}}}},"tags":["Countries"]}}},"info":{"title":"Country API","description":"API для управления странами и валютами","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"Currency":{"type":"object","properties":{"code":{"type":"string","example":"EUR","description":"Код валюты в формате ISO 4217"},"name":{"type":"string","example":"Euro","description":"Название валюты"},"isActive":{"type":"boolean","example":true,"description":"Активна ли валюта для страны"}},"required":["code","name","isActive"]},"Country":{"type":"object","properties":{"name":{"type":"string","example":"France","description":"Название страны"},"isoCode":{"type":"string","example":"FR","description":"ISO-код страны"},"currencies":{"description":"Список валют, используемых в стране","type":"array","items":{"$ref":"#/components/schemas/Currency"}},"isActive":{"type":"boolean","example":true,"description":"Активна ли страна"}},"required":["name","isoCode","currencies","isActive"]},"CreateCountryResponse":{"type":"object","properties":{"status":{"type":"number","example":201},"message":{"type":"string","example":"Страна успешно создана"},"data":{"allOf":[{"$ref":"#/components/schemas/Country"}]}},"required":["status","message","data"]},"GetCountriesResponse":{"type":"object","properties":{"status":{"type":"number","example":200},"message":{"type":"string","example":"Список стран успешно получен"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Country"}}},"required":["status","message","data"]},"UpdateCountryStatusResponse":{"type":"object","properties":{"status":{"type":"number","example":200},"message":{"type":"string","example":"Статус страны обновлён"},"data":{"allOf":[{"$ref":"#/components/schemas/Country"}]}},"required":["status","message","data"]},"UpdateCurrencyStatusResponse":{"type":"object","properties":{"status":{"type":"number","example":200},"message":{"type":"string","example":"Активность валюты успешно изменена"},"data":{"allOf":[{"$ref":"#/components/schemas/Country"}]}},"required":["status","message","data"]},"ValidationErrorResponse":{"type":"object","properties":{"status":{"type":"number","example":400},"message":{"type":"string","example":"Некорректные данные запроса"}},"required":["status","message"]},"ServerErrorResponse":{"type":"object","properties":{"status":{"type":"number","example":500},"message":{"type":"string","example":"Внутренняя ошибка сервера"}},"required":["status","message"]},"ForbiddenErrorResponse":{"type":"object","properties":{"status":{"type":"number","example":403},"message":{"type":"string","example":"Доступ запрещён"}},"required":["status","message"]},"NotFoundErrorResponse":{"type":"object","properties":{"status":{"type":"number","example":404},"message":{"type":"string","example":"Объект не найден"}},"required":["status","message"]},"CurrencyDto":{"type":"object","properties":{"code":{"type":"string","example":"USD","description":"Код валюты (ISO 4217)"},"name":{"type":"string","example":"Доллар США","description":"Название валюты"}},"required":["code","name"]},"CreateCountryDto":{"type":"object","properties":{"name":{"type":"string","example":"США","description":"Название страны"},"isoCode":{"type":"string","example":"US","description":"ISO-код страны"},"currencies":{"type":"array","items":{"$ref":"#/components/schemas/CurrencyDto"}}},"required":["name","isoCode","currencies"]},"UpdateStatusDto":{"type":"object","properties":{"isActive":{"type":"boolean","example":true,"description":"Флаг активности страны"}},"required":["isActive"]}}}}